+++ build-config.ts
@@ -7,13 +7,15 @@ export const webComponentList: WebComponentBuildConfig[] = [
     outputPath: "./web-component/dist/jb-number-input.js",
     tsConfigPath: "./web-component/tsconfig.json",
     umdName: "JBNumberInput",
-    external: ["jb-input", "jb-validation", "jb-core", "jb-core/theme","jb-core/i18n"],
+    external: ["jb-input", "jb-validation", "jb-core", "jb-core/theme", "jb-core/i18n", "jb-icons/minus", "jb-icons/plus"],
     globals: {
       "jb-input": "JBInput",
       "jb-validation": "JBValidation",
-      "jb-core":"JBCore",
-      "jb-core/theme":"JBCoreTheme",
-      "jb-core/i18n": "jb-coreI18N"
+      "jb-core": "JBCore",
+      "jb-core/theme": "JBCoreTheme",
+      "jb-core/i18n": "jb-coreI18N",
+      "jb-icons/minus": "JBIconMinus",
+      "jb-icons/plus": "JBIconPlus",
     },
   },
 ];
@@ -28,10 +30,10 @@ export const reactComponentList: ReactComponentBuildConfig[] = [
       "jb-number-input": "JBNumberInput",
       "jb-input": "JBInput",
       "jb-input/react": "JBInputReact",
-      "jb-core":"JBCore",
-      "jb-core/react":"JBCoreReact"
+      "jb-core": "JBCore",
+      "jb-core/react": "JBCoreReact",
     },
     umdName: "JBNumberReact",
-    dir: "./react"
+    dir: "./react",
   },
 ];